src/forge/pyproject.toml
[project]
name = "forge"
version = "0.2.0"
description = "AI-powered toolkit for open-source maintainers"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Farhan"}]
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2.7.0",
    "pygithub>=2.3.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
    "questionary>=2.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
forge = "forge.cli:main"

[tool.setuptools.packages.find]
where = ["src"]